Add GitHub workflow to generate SunscanOS image
This commit introduces the use of pi-gen to generate the SunscanOS
image. When a tag is created, it will automatically trigger a GitHub
workflow which will generate an image of the system and upload to
the corresponding GitHub release.

In addition, this will create an image everytime something is
merged into the `main` branch.

The name of the image is based on the tag name. If missing, it
will use the commit id.

Building an image takes about 45 minutes. The image is based
on the Raspberry OS Lite image (stage 2).